Remove commented sendToNearby
Although I didn't go through the whole FML chain of classes, the likely
reason why sendToNearby exists is to bypass the whole FML class-chain
to be faster.

However, having another implementation of the same mechanism makes
the code less maintainable, so let's go with the FML one for now.

If there are performance issues due to that, then sure, we can
roll back to the non-FML implementation.